+++ ip2dn.h
@@ -37,7 +37,7 @@ struct map {
 	struct ip2dn *list;
 }__attribute__((packed));
 
-int map_init(struct map *map, uint16_t increm);
+int map_init(struct map *map, uint32_t increm);
 uint32_t hash(uint32_t key, unsigned int cap);
 struct ip2dn *find(struct map *map, uint32_t key);
 int insert(struct map *map, struct ip2dn *entry);
